restructure root navigator
